Refactor index facade and scorer weights
