+++ src/dewi/cli.py
@@ -3,6 +3,7 @@
 import json
 import os
 import sys
+import uuid
 from pathlib import Path
 from typing import List, Optional, Tuple, Dict, Any, cast, Union
